placed show review and author feedback functionality at the right place
